Ive added new files for the the code to look more professional
homepage (which will display motivational pictures i added in the assets),
accountsettings (which will display users information BMI,BMR,AMR, Name, Email,Username,Height,weight,age,gender), and the
calander page (which will display the up to date calander with X's, O's and checkmarks depending on if the gym day was attended/missed/or it was arest day)

there is also the guidelines for the taskbar at the bottom to help navigate through the pages but i havent set inplemented much at the moment. next commit needs to have one of the new pages complete or the task bar completely working.